✨ Fold SubmitOrderItems into CompleteOrder and queue it last
