+++ src/Catalog.ServiceComposition/Workflow/CartWorkflowSlice.cs
@@ -1,11 +1,12 @@
-using Catalog.Endpoint.Messages.Commands;
 using WorkflowComposer;
-using OrderItem = Catalog.Endpoint.Messages.Commands.OrderItem;
 
 namespace Catalog.ServiceComposition.Workflow;
 
-// In-flight cart contents. Stored in checkout.db via WorkflowComposer.
-// Translated into SubmitOrderItems at workflow submit time.
+// In-flight cart contents. Stored in checkout.db via WorkflowComposer
+// and read by the cart/summary composers for UI display. The items
+// are copied into CompleteOrderSlice at submit time, so this slice
+// itself emits no command - CompleteOrder carries the line items
+// directly.
 public sealed record CartSlice(IReadOnlyList<CartLine> Items)
 {
     public static CartSlice Empty { get; } = new([]);
@@ -19,20 +20,5 @@ public class CartWorkflowSlice : WorkflowSlice<CartSlice>
 
     public override string SliceKey => Key;
 
-    protected override object? BuildSubmitCommand(Guid orderId, CartSlice slice)
-    {
-        if (slice.Items.Count == 0) return null;
-
-        return new SubmitOrderItems
-        {
-            OrderId = orderId,
-            Items = slice.Items
-                .Select(i => new OrderItem
-                {
-                    ProductId = i.ProductId,
-                    Quantity = i.Quantity
-                })
-                .ToList()
-        };
-    }
+    protected override object? BuildSubmitCommand(Guid orderId, CartSlice slice) => null;
 }
