+++ src/WorkflowComposer/IWorkflowSlice.cs
@@ -13,6 +13,12 @@ public interface IWorkflowSlice
     // The CLR type the framework should (de)serialize for this slice.
     Type SliceType { get; }
 
+    // Determines the order in which slices' commands are enqueued at
+    // submit time. Lower values go first; ties keep DI order. Use a
+    // high value for finalizer-style slices (e.g. CompleteOrder) so
+    // the per-boundary writes are queued before the trigger.
+    int SubmitOrder => 0;
+
     // Translate the persisted slice into the message that gets sent
     // to the boundary's endpoint when the workflow submits. Return
     // null to skip (e.g., the slice was never written).
@@ -27,6 +33,8 @@ public abstract class WorkflowSlice<TSlice> : IWorkflowSlice
     public abstract string SliceKey { get; }
 
     public Type SliceType => typeof(TSlice);
+
+    public virtual int SubmitOrder => 0;
 
     public object? BuildSubmitCommand(Guid workflowId, object slice) =>
         BuildSubmitCommand(workflowId, (TSlice)slice);
